Update building_rest_apis.md
The documented way did not work for me. Using that documented way, when I accessed GET /api/user  the index fn ran but when I accessed GET /api/user/4   the index fn ran not the view fn.  I got it to work by reversing the order of the routes so the more narrow route was listed first and removing the ? in the pattern. If I didn't remove the ? in the pattern but still listed that first, both /api/user  and /api/user/3  ran the view fn.