Update README.md
In my brain I always call a requestContext object the 'event'. I always forget that argument 'event' for an action is of type coldbox.system.web.context.RequestContext. 

This is to save other people like me having to google the docs for 30s instead of 2-3 minutes just to read the API docs