Update request-context.md
In my brain I always call a RequestContext object the 'event'. I always forget that the argument 'event' for an action is of type coldbox.system.web.context.RequestContext. 

This is to save other people like me having to search the docs for 30s instead of 2-3 minutes just remember this